perf(search): add native workspace search
