+++ apps/desktop/src-tauri/src/lib.rs
@@ -23,7 +23,8 @@ use markdown_files::{
     move_markdown_tree_file, open_markdown_file_in_new_window, open_markdown_folder_in_new_window,
     open_markdown_path, read_markdown_file, read_markdown_file_history, read_markdown_image_file,
     read_markdown_template_file, rename_markdown_tree_file, resolve_markdown_path,
-    save_clipboard_image, write_markdown_file, write_markdown_template_file,
+    save_clipboard_image, search_markdown_files_for_path, write_markdown_file,
+    write_markdown_template_file,
 };
 use menu::{
     apply_native_application_menu_for_window_event, create_application_menu,
@@ -110,6 +111,7 @@ pub fn run() {
         })
         .invoke_handler(tauri::generate_handler![
             list_markdown_files_for_path,
+            search_markdown_files_for_path,
             create_markdown_tree_file,
             create_markdown_tree_folder,
             install_application_menu,
