+++ packages/app/src/lib/workspace-search.ts
@@ -17,7 +17,20 @@ export type WorkspaceSearchResult = {
 export type WorkspaceSearchResponse = {
   results: WorkspaceSearchResult[];
   searchedFileCount: number;
+  truncated: boolean;
   unreadableFileCount: number;
+};
+
+export type WorkspaceSearchRequest = {
+  caseSensitive?: boolean;
+  currentDocument?: {
+    content: string;
+    path: string;
+  } | null;
+  maxMatches?: number;
+  maxMatchesPerFile?: number;
+  path: string;
+  query: string;
 };
 
 type WorkspaceSearchReadResult = {
@@ -54,6 +67,7 @@ export async function searchWorkspaceFiles(
     return {
       results: [],
       searchedFileCount: searchableFiles.length,
+      truncated: false,
       unreadableFileCount: 0
     };
   }
@@ -65,7 +79,7 @@ export async function searchWorkspaceFiles(
 
         return {
           file,
-          matches: findWorkspaceSearchResults(file, read.content, normalizedQuery, {
+          ...findWorkspaceSearchResults(file, read.content, normalizedQuery, {
             caseSensitive: options.caseSensitive,
             maxMatchesPerFile
           }),
@@ -75,6 +89,7 @@ export async function searchWorkspaceFiles(
         return {
           file,
           matches: [] as WorkspaceSearchResult[],
+          truncated: false,
           unreadable: true
         };
       }
@@ -83,6 +98,8 @@ export async function searchWorkspaceFiles(
 
   const results: WorkspaceSearchResult[] = [];
   const unreadableFileCount = searched.filter((item) => item.unreadable).length;
+  const collectedMatchCount = searched.reduce((count, item) => count + item.matches.length, 0);
+  const truncatedByFileLimit = searched.some((item) => item.truncated);
 
   for (const item of searched) {
     for (const match of item.matches) {
@@ -96,6 +113,7 @@ export async function searchWorkspaceFiles(
   return {
     results,
     searchedFileCount: searchableFiles.length,
+    truncated: truncatedByFileLimit || collectedMatchCount > maxMatches,
     unreadableFileCount
   };
 }
@@ -106,11 +124,14 @@ function findWorkspaceSearchResults(
   query: string,
   options: { caseSensitive?: boolean; maxMatchesPerFile: number }
 ) {
+  const searchLimit = options.maxMatchesPerFile + 1;
   const ranges = findSearchRanges(content, query, {
-    caseSensitive: options.caseSensitive
-  }).slice(0, options.maxMatchesPerFile);
+    caseSensitive: options.caseSensitive,
+    maxMatches: searchLimit
+  });
+  const truncated = ranges.length > options.maxMatchesPerFile;
 
-  return ranges.map((range, matchIndex) => {
+  const results = ranges.slice(0, options.maxMatchesPerFile).map((range, matchIndex) => {
     const line = lineForSearchRange(content, range);
 
     return {
@@ -124,6 +145,11 @@ function findWorkspaceSearchResults(
       snippet: workspaceSearchSnippet(line.text, line.columnNumber, range.to - range.from)
     } satisfies WorkspaceSearchResult;
   });
+
+  return {
+    matches: results,
+    truncated
+  };
 }
 
 function lineForSearchRange(content: string, range: SearchRange) {
